requirements.txt
jupyterlab==3.4.4

mlxtend==0.20.0

openpyxl==3.0.10

pandas==1.4.3

seaborn==0.11.2

scikit-learn==1.1.1

gensim==4.2.0

spacy==3.4.0

tqdm==4.64.0

durable-rules==2.0.28

nltk==3.7

pyro-ppl==1.8.1

torch==1.12.0

torchvision==0.13.0